r5.2.2

+++ shaders/lib/pipelineSettings.glsl
@@ -1,7 +1,7 @@
 /*
 const int colortex0Format = R11F_G11F_B10F; //main color
 
-const int colortex2Format = RGBA16;         //taa & previous depth
+const int colortex2Format = RGBA16F;        //taa & previous depth
 const int colortex3Format = RGBA8;          //(cloud/water map on deferred/gbuffer) | translucentMult & bloom & final color
 const int colortex4Format = R8;             //volumetric cloud linear depth & volumetric light factor
 const int colortex5Format = RGBA8_SNORM;    //normalM & scene image for water reflections
